[PR 8223] [vfs] Fix crash when entering zip:// URLs with unknown zip file size
For example, consider an archive hosted at the URL:

https://test.com/archive.zip

The zip:// URL for a file in the archive might look like:

zip://https%3A%2F%2Ftest.com%2Farchive.zip/file.txt

In this case, a CURL file is used to load the .zip. However, the size of the
remote .zip is initially unknown (GetLength() returns 0), causing the crash.