[retroplayer] Add window mapping 'fullscreengame'
This commit adds a window mapping entry for 'fullscreengame' to the
ButtonTranslator so that keymaps can be written for full screen gaming.

A fallback to 'fullscreenvideo' ensures that when the keymap does not contain
a specific section for 'fullscreengame', it uses the 'fullscreenvideo' mapping.
This way without additional configuration, a remote can be used in parallel
with the gamepad for example to show the OSD menu.