mmal: Switch to a pool class attached to decoder output
Pool containing mmal buffers (e.g. decoded pictures) need to be attached
to a mmal components port.

Managing the lifespan of these is tricky, as renderer may consitue to
display an image after the decoder has been closed and a new one opened.

So create a class for handling this. The class creates and destroys the
mmal component associated with the pool, and creates the mmal pool
(which is just headers) as well as handling allocation of the gpu buffers.

Each object from the pool includes a shared pointer to the pool class,
ensuring the pool and mmal component are not destroyed until the last
object is destroyed.

The pool class is used by both MMAL hw decode and the FFMPEG zero copy
decoders which avoids some special casing in the renderer.

It will also handle the pool of deinterlaced frames later.