[PR 10630] [joysticks] Only dispatch action if button was pressed this frame
This fixes a future problem when the "cheat code" to enable games
is merged. The cheat code is handled in the current frame, but in
the next frame, the button's Kodi action is executed.

This fixes the problem by only executing a Kodi action if the
button hasn't been held for more than 1 frame (hold time > 0).