[PR 11571] [peripheral API] v1.3.0 - Add handling for anomalous triggers
Anomalous triggers are triggers with a non-zero offset. For example, if a
trigger's rest position is +1 and it travels to -1 when fully pressed, then
it is an anomalous trigger with a center of +1 and a range of 2.

Two properties are added to the definition of a semiaxis:

* Center
* Range

The center is the rest position of the trigger. The range is 1 if the
trigger stops at 0, or 2 if the trigger continues past zero to the opposite
end of the axis.